+++ apps/veil/src/pages/tournament/api/use-voting-notes.ts
@@ -6,7 +6,7 @@ import { useRefetchOnNewBlock } from '@/shared/api/compact-block';
 import { connectionStore } from '@/shared/model/connection';
 import { penumbra } from '@/shared/const/penumbra';
 
-const fetchQuery = async (subaccount = 0, epoch: number): Promise<LqtVotingNotesResponse[]> => {
+const fetchQuery = async (subaccount: number, epoch: number): Promise<LqtVotingNotesResponse[]> => {
   const accountFilter = new AddressIndex({ account: subaccount });
 
   return Array.fromAsync(
@@ -17,16 +17,16 @@ const fetchQuery = async (subaccount = 0, epoch: number): Promise<LqtVotingNotes
 /**
  * Must be used within the `observer` mobX HOC
  */
-export const useLQTNotes = (subaccount = 0, epoch?: number) => {
+export const useLQTNotes = (subaccount: number, epoch?: number, disabled?: boolean) => {
   const lqtNotesQuery = useQuery({
     queryKey: ['lqt-notes', subaccount, epoch],
     staleTime: Infinity,
-    enabled: connectionStore.connected && !!epoch,
+    enabled: connectionStore.connected && !!epoch && !disabled,
     // eslint-disable-next-line @typescript-eslint/no-non-null-assertion -- based on `enabled`, epoch is always defined
     queryFn: () => fetchQuery(subaccount, epoch!),
   });
 
-  useRefetchOnNewBlock('lqt-notes', lqtNotesQuery);
+  useRefetchOnNewBlock('lqt-notes', lqtNotesQuery, disabled);
 
   return lqtNotesQuery;
 };