+++ apps/veil/src/pages/tournament/server/summary.ts
@@ -1,11 +1,12 @@
 import { NextRequest, NextResponse } from 'next/server';
 import { Serialized } from '@/shared/utils/serializer';
 import { pindexerDb } from '@/shared/database/client';
-import { LQTSummary } from '@/shared/database/schema';
+import { LqtSummary } from '@/shared/database/schema';
 
 export interface TournamentSummaryRequest {
   limit: number;
   page: number;
+  epoch?: number;
 }
 
 const DEFAULT_LIMIT = 10;
@@ -16,23 +17,27 @@ export const getQueryParams = (req: NextRequest): TournamentSummaryRequest => {
   const limit = Number(searchParams.get('limit')) || DEFAULT_LIMIT;
   const page = Number(searchParams.get('page')) || 1;
 
+  const epoch = Number(searchParams.get('epoch')) || undefined;
+
   return {
+    epoch,
     limit,
     page,
   };
 };
 
 export interface TournamentSummaryApiResponse {
-  data: LQTSummary[];
+  data: LqtSummary[];
 }
 
-const tournamentSummaryQuery = async ({ limit, page }: TournamentSummaryRequest) => {
+const tournamentSummaryQuery = async ({ limit, page, epoch }: TournamentSummaryRequest) => {
   return pindexerDb
     .selectFrom('lqt.summary')
     .orderBy('epoch', 'desc')
     .limit(limit)
     .offset(limit * (page - 1))
     .selectAll()
+    .$if(!!epoch, qb => (epoch ? qb.where('epoch', '=', epoch) : qb))
     .execute();
 };