+++ apps/veil/src/pages/tournament/ui/page.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { PenumbraWaves } from '@/pages/explore/ui/waves';
-import { MyRewards } from './my-rewards';
+import { DelegatorRewards } from './delegator-rewards';
 import { LandingCard } from './landing-card';
 import { DelegatorLeaderboard } from './delegator-leaderboard';
 import { PreviousEpochs } from './previous-epochs';
@@ -11,7 +11,7 @@ export const TournamentPage = () => {
     <section className='flex flex-col gap-6 p-4 max-w-[1168px] mx-auto'>
       <PenumbraWaves />
       <LandingCard />
-      <MyRewards />
+      <DelegatorRewards />
 
       <DelegatorLeaderboard />
       <PreviousEpochs />