+++ apps/veil/src/pages/tournament/ui/previous-epochs.tsx
@@ -1,6 +1,6 @@
 import cn from 'clsx';
 import Link from 'next/link';
-import { useMemo, useState } from 'react';
+import { useState } from 'react';
 import { observer } from 'mobx-react-lite';
 import { ChevronRight } from 'lucide-react';
 import { ValueView } from '@penumbra-zone/protobuf/penumbra/core/asset/v1/asset_pb';
@@ -12,12 +12,13 @@ import { Density } from '@penumbra-zone/ui/Density';
 import { Button } from '@penumbra-zone/ui/Button';
 import { Text } from '@penumbra-zone/ui/Text';
 import { connectionStore } from '@/shared/model/connection';
+import { useStakingTokenMetadata } from '@/shared/api/registry';
 import { usePreviousEpochs, BASE_PAGE, BASE_LIMIT } from '../api/use-previous-epochs';
 import type { PreviousEpochData } from '../server/previous-epochs';
 import { useSortableTableHeaders } from './sortable-table-header';
 import { usePersonalRewards } from '../api/use-personal-rewards';
-import { GENERIC_DELUM } from './shared/delum-metadata';
 import { Vote } from './vote';
+import { pnum } from '@penumbra-zone/types/pnum';
 
 const TABLE_CLASSES = {
   table: {
@@ -37,79 +38,82 @@ interface PreviousEpochsRowProps {
   connected: boolean;
 }
 
-const PreviousEpochsRow = ({ row, isLoading, className, connected }: PreviousEpochsRowProps) => {
-  const { data: rewards, isLoading: rewardsLoading } = usePersonalRewards(row.epoch);
+const PreviousEpochsRow = observer(
+  ({ row, isLoading, className, connected }: PreviousEpochsRowProps) => {
+    const { subaccount } = connectionStore;
+    const { data: rewards, isLoading: rewardsLoading } = usePersonalRewards(subaccount, row.epoch);
+    const { data: stakingToken } = useStakingTokenMetadata();
 
-  const votingRewards = useMemo(() => {
-    const reward = rewards?.[0];
-    if (!reward?.reward) {
-      return '-';
-    }
-
-    const value = new ValueView({
-      valueView: {
-        case: 'knownAssetId',
-        value: {
-          amount: reward.reward.amount,
-          metadata: GENERIC_DELUM,
-        },
-      },
-    });
-
-    return <ValueViewComponent valueView={value} priority='tertiary' />;
-  }, [rewards]);
-
-  return (
-    <Link
-      href={isLoading ? '' : `/tournament/${row.epoch}`}
-      className={cn(
-        className,
-        'grid grid-cols-subgrid',
-        'hover:bg-action-hoverOverlay transition-colors cursor-pointer',
-      )}
-    >
-      <TableCell cell loading={isLoading}>
-        Epoch #{row.epoch}
-      </TableCell>
-      <TableCell cell loading={isLoading}>
-        {!isLoading && !row.gauge.length && '-'}
-        {!isLoading &&
-          row.gauge.slice(0, 3).map((vote, index) => (
-            <div key={index} className='flex items-center justify-start min-w-[88px] px-1'>
-              <Vote asset={vote.asset} percent={vote.portion} hideFor />
-            </div>
-          ))}
-        {!isLoading && row.gauge.length > 3 && (
-          <Tooltip
-            message={
-              <div className='flex flex-col gap-2'>
-                {row.gauge.slice(3).map((vote, index) => (
-                  <Vote key={index} asset={vote.asset} percent={vote.portion} hideFor />
-                ))}
-              </div>
-            }
-          >
-            <div className='flex items-center justify-start px-3 text-text-primary'>
-              <Text smallTechnical>+{row.gauge.length - 3}</Text>
-            </div>
-          </Tooltip>
+    return (
+      <Link
+        href={isLoading ? '' : `/tournament/${row.epoch}`}
+        className={cn(
+          className,
+          'grid grid-cols-subgrid',
+          'hover:bg-action-hoverOverlay transition-colors cursor-pointer',
         )}
-      </TableCell>
-      {connected && (
-        <TableCell cell loading={isLoading || rewardsLoading}>
-          {votingRewards}
+      >
+        <TableCell cell loading={isLoading}>
+          Epoch #{row.epoch}
         </TableCell>
-      )}
-      <TableCell cell loading={isLoading}>
-        <Density slim>
-          <Button iconOnly icon={ChevronRight}>
-            Go to voting reward information
-          </Button>
-        </Density>
-      </TableCell>
-    </Link>
-  );
-};
+        <TableCell cell loading={isLoading}>
+          {!isLoading && !row.gauge.length && '-'}
+          {!isLoading &&
+            row.gauge.slice(0, 3).map((vote, index) => (
+              <div key={index} className='flex items-center justify-start min-w-[88px] px-1'>
+                <Vote asset={vote.asset} percent={vote.portion} hideFor />
+              </div>
+            ))}
+          {!isLoading && row.gauge.length > 3 && (
+            <Tooltip
+              message={
+                <div className='flex flex-col gap-2'>
+                  {row.gauge.slice(3).map((vote, index) => (
+                    <Vote key={index} asset={vote.asset} percent={vote.portion} hideFor />
+                  ))}
+                </div>
+              }
+            >
+              <div className='flex items-center justify-start px-3 text-text-primary'>
+                <Text smallTechnical>+{row.gauge.length - 3}</Text>
+              </div>
+            </Tooltip>
+          )}
+        </TableCell>
+        {connected && (
+          <TableCell cell loading={isLoading || rewardsLoading}>
+            {rewards?.data.find(r => BigInt(r.epoch) === (row.epoch && BigInt(row.epoch))) &&
+              stakingToken && (
+                <ValueViewComponent
+                  valueView={
+                    new ValueView({
+                      valueView: {
+                        case: 'knownAssetId',
+                        value: {
+                          amount: pnum(
+                            rewards.data.find(r => BigInt(r.epoch) === BigInt(row.epoch))?.reward,
+                          ).toAmount(),
+                          metadata: stakingToken,
+                        },
+                      },
+                    })
+                  }
+                  priority='tertiary'
+                />
+              )}
+          </TableCell>
+        )}
+        <TableCell cell loading={isLoading}>
+          <Density slim>
+            <Button iconOnly icon={ChevronRight}>
+              Go to voting reward information
+            </Button>
+          </Density>
+        </TableCell>
+      </Link>
+    );
+  },
+);
 
 export const PreviousEpochs = observer(() => {
   const [page, setPage] = useState(BASE_PAGE);
@@ -147,10 +151,10 @@ export const PreviousEpochs = observer(() => {
             <TableCell heading> </TableCell>
           </div>
 
-          {epochs.map(epoch => (
+          {epochs.map((epoch, index) => (
             <PreviousEpochsRow
-              key={epoch.epoch}
-              isLoading={isLoading || !epoch.gauge}
+              key={isLoading ? index : epoch.epoch}
+              isLoading={isLoading}
               row={epoch}
               connected={connected}
               className={TABLE_CLASSES.row[tableKey]}