+++ apps/veil/src/pages/tournament/ui/round/page.tsx
@@ -12,9 +12,6 @@ export const TournamentRoundPage = () => {
   const router = useRouter();
   const epoch = Number(params?.epoch);
 
-  const startBlock = 1337;
-  const endBlock = 1337;
-
   if (!params?.epoch && !Number.isNaN(epoch)) {
     router.push(PagePath.Tournament);
     return null;
@@ -24,8 +21,8 @@ export const TournamentRoundPage = () => {
     <section className='flex flex-col gap-6 p-4 max-w-[1168px] mx-auto'>
       <PenumbraWaves />
       <RoundCard epoch={epoch} />
-      <CurrentVotingResults />
-      <LeaderboardTable startBlock={startBlock} endBlock={endBlock} />
+      <CurrentVotingResults epoch={epoch} />
+      <LeaderboardTable />
     </section>
   );
 };