+++ apps/veil/src/pages/tournament/ui/vote-dialog/index.tsx
@@ -4,6 +4,7 @@ import { observer } from 'mobx-react-lite';
 import { Dialog } from '@penumbra-zone/ui/Dialog';
 import { Text } from '@penumbra-zone/ui/Text';
 import { Button } from '@penumbra-zone/ui/Button';
+import { Density } from '@penumbra-zone/ui/Density';
 import { Checkbox } from '@penumbra-zone/ui/Checkbox';
 import { TextInput } from '@penumbra-zone/ui/TextInput';
 import { SpendableNoteRecord } from '@penumbra-zone/protobuf/penumbra/view/v1/view_pb';
@@ -11,15 +12,15 @@ import { useSubaccounts } from '@/widgets/header/api/subaccounts';
 import { connectionStore } from '@/shared/model/connection';
 import { getAddressIndex } from '@penumbra-zone/getters/address-view';
 import { useLQTNotes } from '../../api/use-voting-notes';
-import { voteTournament } from '../../api/vote';
+import { voteTournament } from '../../../../entities/tournament/api/vote';
 import { useCurrentEpoch } from '../../api/use-current-epoch';
-import { useEpochGauge } from '../../api/use-epoch-gauge';
+import { useEpochResults } from '../../api/use-epoch-results';
 import { MappedGauge } from '../../server/previous-epochs';
 import { VoteDialogSearchResults } from './search-results';
 import { VotingAssetSelector } from './asset-selector';
 
 interface VoteDialogProps {
-  defaultValue?: string;
+  defaultValue?: MappedGauge;
   isOpen: boolean;
   onClose: () => void;
 }
@@ -28,8 +29,7 @@ const DEFAULT_REVEAL_VOTE = true;
 
 export const VoteDialogueSelector = observer(
   ({ isOpen, onClose, defaultValue }: VoteDialogProps) => {
-    const [selectedDenom, setSelectedDenom] = useState<string | undefined>(defaultValue);
-    const [selectedAsset, setSelectedAsset] = useState<MappedGauge | undefined>();
+    const [selectedAsset, setSelectedAsset] = useState<MappedGauge | undefined>(defaultValue);
     const [revealVote, setRevealVote] = useState(DEFAULT_REVEAL_VOTE);
 
     const [isSearchOpen, setIsSearchOpen] = useState(false);
@@ -42,42 +42,51 @@ export const VoteDialogueSelector = observer(
       account => getAddressIndex(account).account === subaccount,
     );
 
+    // Temporarily hardcode the same account address as the reward recipient.
     const rewardsRecipient = valueAddress?.addressView.value?.address;
 
     // Fetch user's spendable voting notes for this epoch
-    const { epoch } = useCurrentEpoch();
-    const { data: notes } = useLQTNotes(subaccount, epoch);
+    const { epoch, isLoading: epochLoading } = useCurrentEpoch();
 
-    const { data: assets, isLoading } = useEpochGauge(epoch);
-    const isEmptyScreen = !isLoading && !!assets && !assets.length;
+    const { data: notes } = useLQTNotes(subaccount, epoch);
+    const { data: assets, isLoading } = useEpochResults(
+      'epoch-results-vote-dialog',
+      {
+        epoch,
+        limit: 30,
+        page: 1,
+      },
+      !isOpen && epochLoading,
+    );
 
     const handleVoteSubmit = async () => {
-      if (selectedDenom) {
-        if (!epoch) {
-          throw new Error('Missing epoch index');
-        }
-
-        const stakedNotes: SpendableNoteRecord[] = notes
-          ? Array.from(notes.values())
-              .map(res => res.noteRecord)
-              .filter((record): record is SpendableNoteRecord => !!record)
-          : [];
-
-        // Craft LQT TPR and submit vote
-        await voteTournament({
-          stakedNotes: stakedNotes,
-          incentivized: selectedDenom,
-          epochIndex: epoch,
-          rewardsRecipient,
-        });
-
-        handleClose();
+      if (!selectedAsset) {
+        throw new Error('Please, select an asset to vote for');
       }
+
+      if (!epoch) {
+        throw new Error('Missing epoch index');
+      }
+
+      const stakedNotes: SpendableNoteRecord[] = notes
+        ? Array.from(notes.values())
+            .map(res => res.noteRecord)
+            .filter((record): record is SpendableNoteRecord => !!record)
+        : [];
+
+      // Craft LQT TPR and submit vote
+      await voteTournament({
+        stakedNotes: stakedNotes,
+        incentivized: selectedAsset.asset.base,
+        epochIndex: epoch,
+        rewardsRecipient,
+      });
+
+      handleClose();
     };
 
     const handleClose = () => {
       setSearchQuery('');
-      setSelectedDenom(undefined);
       setSelectedAsset(undefined);
       setIsSearchOpen(false);
       setRevealVote(DEFAULT_REVEAL_VOTE);
@@ -87,7 +96,6 @@ export const VoteDialogueSelector = observer(
     const onSearchSelect = (asset: MappedGauge) => {
       setSearchQuery('');
       setIsSearchOpen(false);
-      setSelectedDenom(asset.asset.base);
       setSelectedAsset(asset);
     };
 
@@ -121,20 +129,22 @@ export const VoteDialogueSelector = observer(
           }
           buttons={
             !isSearchOpen && (
-              <div className='flex flex-col gap-6 px-6 pb-6 pt-2 [&>label]:justify-center [&>label>div]:grow-0'>
-                <Button
-                  onClick={() => {
-                    onClose();
-                    handleVoteSubmit().catch((err: unknown) => console.error(err));
-                  }}
-                  priority='primary'
-                  actionType='accent'
-                  disabled={!selectedDenom}
-                >
-                  {selectedDenom
-                    ? `Vote for ${selectedAsset?.asset.symbol ?? selectedDenom.toUpperCase()}`
-                    : 'Select asset to vote'}
-                </Button>
+              <div className='flex flex-col gap-6 px-6 pb-6 [&>label]:justify-center [&>label>div]:grow-0'>
+                <Density sparse>
+                  <Button
+                    onClick={() => {
+                      onClose();
+                      handleVoteSubmit().catch((err: unknown) => console.error(err));
+                    }}
+                    priority='primary'
+                    actionType='accent'
+                    disabled={!selectedAsset}
+                  >
+                    {selectedAsset
+                      ? `Vote for ${selectedAsset.asset.symbol}`
+                      : 'Select asset to vote'}
+                  </Button>
+                </Density>
 
                 <Checkbox
                   title='Reveal my vote to the leaderboard.'
@@ -147,20 +157,19 @@ export const VoteDialogueSelector = observer(
           }
         >
           <div className='flex flex-col pt-2'>
-            {!isSearchOpen && !isEmptyScreen && (
+            {!isSearchOpen && (
               <VotingAssetSelector
-                value={selectedDenom}
                 selectedAsset={selectedAsset}
                 loading={isLoading}
-                gauge={assets ?? []}
+                gauge={assets?.data ?? []}
                 onSelect={onSearchSelect}
               />
             )}
 
-            {(isSearchOpen || isEmptyScreen) && (
+            {isSearchOpen && (
               <VoteDialogSearchResults
-                value={selectedDenom}
-                gauge={assets ?? []}
+                value={selectedAsset?.asset.base}
+                gauge={assets?.data ?? []}
                 search={searchQuery}
                 onSelect={onSearchSelect}
               />