+++ packages/storage/src/indexed-db/index.ts
@@ -428,13 +428,18 @@ export class IndexedDb implements IndexedDbInterface {
   /**
    * Retrieves liquidity tournament votes and rewards for a given epoch.
    */
-  async getLQTHistoricalVotes(epoch: bigint): Promise<
+  async getLQTHistoricalVotes(
+    epoch: bigint,
+    subaccount?: number,
+  ): Promise<
     {
+      epoch: string;
       TransactionId: TransactionId;
       AssetMetadata: Metadata;
       VoteValue: Value;
       RewardValue: Amount | undefined;
       id: string | undefined;
+      subaccount?: number;
     }[]
   > {
     const tournamentVotes = await this.db.getAllFromIndex(
@@ -443,7 +448,12 @@ export class IndexedDb implements IndexedDbInterface {
       epoch.toString(),
     );
 
-    return tournamentVotes.map(tournamentVote => ({
+    const filtered = subaccount
+      ? tournamentVotes.filter(vote => vote.subaccount === subaccount)
+      : tournamentVotes;
+
+    return filtered.map(tournamentVote => ({
+      epoch: tournamentVote.epoch,
       TransactionId: TransactionId.fromJson(tournamentVote.TransactionId, { typeRegistry }),
       AssetMetadata: Metadata.fromJson(tournamentVote.AssetMetadata, { typeRegistry }),
       VoteValue: Value.fromJson(tournamentVote.VoteValue, { typeRegistry }),
@@ -451,7 +461,46 @@ export class IndexedDb implements IndexedDbInterface {
         ? Amount.fromJson(tournamentVote.RewardValue, { typeRegistry })
         : undefined,
       id: tournamentVote.id,
+      subaccount: tournamentVote.subaccount,
     }));
+  }
+
+  /**
+   * Streams tournament votes up to the specified epoch (exclusive),
+   * filtering only those that have a reward.
+   */
+  async *iterateLQTVotes(epoch: bigint, subaccount?: number) {
+    yield* new ReadableStream({
+      start: async cont => {
+        let cursor = await this.db
+          .transaction('LQT_HISTORICAL_VOTES')
+          .store.index('epoch')
+          .openCursor();
+
+        while (cursor) {
+          const voteEpoch = BigInt(cursor.value.epoch);
+          const voteSubaccount = cursor.value.subaccount;
+
+          if (voteEpoch < epoch && voteSubaccount === subaccount) {
+            if (cursor.value.RewardValue) {
+              cont.enqueue({
+                epoch: cursor.value.epoch,
+                TransactionId: TransactionId.fromJson(cursor.value.TransactionId, { typeRegistry }),
+                AssetMetadata: Metadata.fromJson(cursor.value.AssetMetadata, { typeRegistry }),
+                VoteValue: Value.fromJson(cursor.value.VoteValue, { typeRegistry }),
+                RewardValue: Amount.fromJson(cursor.value.RewardValue, { typeRegistry }),
+                id: cursor.value.id,
+                subaccount: cursor.value.subaccount,
+              });
+            }
+          }
+
+          cursor = await cursor.continue();
+        }
+
+        cont.close();
+      },
+    });
   }
 
   /**
@@ -464,6 +513,7 @@ export class IndexedDb implements IndexedDbInterface {
     voteValue: Value,
     rewardValue?: Amount,
     id?: string,
+    subaccount?: number,
   ): Promise<void> {
     assertTransactionId(transactionId);
 
@@ -471,6 +521,8 @@ export class IndexedDb implements IndexedDbInterface {
     // a random one is generated and used to store an object.
     const uniquePrimaryKey = id ?? crypto.randomUUID();
 
+    // TODO: saving the entire metadata is extraneous, experiment with changing
+    // @see https://github.com/penumbra-zone/web/issues/2032.
     const tournamentVote = {
       epoch: epoch.toString(),
       TransactionId: transactionId.toJson({ typeRegistry }) as Jsonified<TransactionId>,
@@ -478,6 +530,7 @@ export class IndexedDb implements IndexedDbInterface {
       VoteValue: voteValue.toJson({ typeRegistry }) as Jsonified<Value>,
       RewardValue: rewardValue ? (rewardValue.toJson({ typeRegistry }) as Jsonified<Amount>) : null,
       id: uniquePrimaryKey,
+      subaccount,
     };
 
     await this.u.update({