+++ apps/veil/CHANGELOG.md
@@ -1,5 +1,33 @@
 # penumbra-veil
 
+## 0.1.0
+
+### Minor Changes
+
+- 694319c: bump registry version
+
+### Patch Changes
+
+- Updated dependencies [4a51a46]
+- Updated dependencies [4a51a46]
+  - @penumbra-zone/ui@14.0.0
+
+## 0.0.9
+
+### Patch Changes
+
+- Updated dependencies [93f1d05]
+- Updated dependencies [28a251c]
+  - @penumbra-zone/protobuf@10.0.0
+  - @penumbra-zone/types@33.0.0
+  - @penumbra-zone/wasm@46.0.0
+  - @penumbra-zone/ui@13.18.0
+  - @penumbra-zone/bech32m@16.0.0
+  - @penumbra-zone/client@27.0.0
+  - @penumbra-zone/getters@26.0.0
+  - @penumbra-zone/perspective@55.0.0
+  - @penumbra-zone/crypto-web@42.0.0
+
 ## 0.0.8
 
 ### Patch Changes