+++ packages/ui-deprecated/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @penumbra-zone/ui
 
+## 19.0.0
+
+### Major Changes
+
+- 4a51a46: Upgrade to React 19
+
+## 18.0.0
+
+### Patch Changes
+
+- Updated dependencies [93f1d05]
+  - @penumbra-zone/protobuf@10.0.0
+  - @penumbra-zone/types@33.0.0
+  - @penumbra-zone/bech32m@16.0.0
+  - @penumbra-zone/perspective@55.0.0
+
 ## 17.1.6
 
 ### Patch Changes